TeamAnalyzer: Updated to 2.70, changed "Diff%" to "Relative" and using (myRating/(myRating+opponentRating)) as new formula, closes #78
git-svn-id: file:///home/coren/project/backup_ho/trunk/Core@517 b8de521a-fa13-0410-8113-fc7ffabbd59a

+++ src/java/de/hattrickorganizer/model/HOVerwaltung.java
@@ -76,7 +76,7 @@ public class HOVerwaltung {
      *
      * @return average TSI
      */
-    public float getAvgTSI() {
+    public double getAvgTSI() {
         int numPlayers = getModel().getAllSpieler().size();
         //Trainer abziehen // without trainer
         if (numPlayers <= 1)
@@ -91,7 +91,7 @@ public class HOVerwaltung {
      *
      * @return average EPV
      */
-    public float getAvgEPV() {
+    public double getAvgEPV() {
         int numPlayers = getModel().getAllSpieler().size();
         //Trainer abziehen // without trainer
         if (numPlayers <= 1)
@@ -105,7 +105,7 @@ public class HOVerwaltung {
      *
      * @return TODO Missing Return Method Documentation
      */
-    public float getDurchschnittsAlter() {
+    public double getDurchschnittsAlter() {
         float summe = 0;
         final Vector<ISpieler> vSpieler = getModel().getAllSpieler();
